Move helpers to namespace root
- Move `RuntimeIdentifier` to its own type, not nested.
- Move helper code out of `Upgraders` namespace.
