+++ src/DotNetBumper.Core/PathHelpers.cs
@@ -4,7 +4,7 @@
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 
-namespace MartinCostello.DotNetBumper.Upgraders;
+namespace MartinCostello.DotNetBumper;
 
 internal static class PathHelpers
 {
