+++ src/DotNetBumper.Core/PostProcessors/LeftoverReferencesPostProcessor.cs
@@ -4,7 +4,6 @@
 using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
 using MartinCostello.DotNetBumper.Logging;
-using MartinCostello.DotNetBumper.Upgraders;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Spectre.Console;
@@ -46,7 +45,7 @@ internal sealed partial class LeftoverReferencesPostProcessor(
 
             if (channel >= DotNetVersions.EightPointZero)
             {
-                matches = line.MatchRuntimeIdentifiers();
+                matches = RuntimeIdentifier.Match(line);
 
                 foreach (var match in matches)
                 {
