+++ src/DotNetBumper.Core/TargetFrameworkHelpers.cs
